Finished newsletter endpoints including errors

+++ newsletter/routes/subscriptionRoutes.js
@@ -1,21 +1,43 @@
 const express = require("express");
 const router = express.Router();
+const { BadRequest } = require("../../api/utils/errors");
+const NonexistentResourceError = require("../../error/nonexistentResourceError");
 
 // Obtener todos los emails suscritos a un artista
-router.get("/", (req, res) => {
+router.get("/", (req, res, next) => {
+  const body = req.body;
   const artistId = parseInt(req.query.artistId);
-  const subscriptors = req.body.observerManager.getSubscriptions(artistId);
-  res.send({
-    artistId: artistId,
-    subscriptors: subscriptors,
+  let subscriptors;
+  if (!artistId) return next(new BadRequest());
+
+  body.unqfyClient.existsArtist(artistId).then((exists) => {
+    if (exists) subscriptors = body.observerManager.getSubscriptions(artistId);
+    else
+      return next(
+        new NonexistentResourceError("UNQfy", artistId, "GetArtist")
+      );
+
+    res.send({
+      artistId: artistId,
+      subscriptors: subscriptors,
+    });
   });
 });
 
 // Eliminar todos los emails suscritos a un artista
-router.delete("/", (req, res) => {
-    const artistId = parseInt(req.body.artistId);
-    req.body.observerManager.deleteSubscriptions(artistId);
+router.delete("/", (req, res, next) => {
+  const artistId = parseInt(req.body.artistId);
+  if (!artistId) return next(new BadRequest());
+
+  req.body.unqfyClient.existsArtist(artistId).then((exists) => {
+    if (exists) req.body.observerManager.deleteSubscriptions(artistId);
+    else
+      return next(
+        new NonexistentResourceError("UNQfy", artistId, "GetArtist")
+      );
+
     res.send();
   });
+});
 
 module.exports = router;
